requirements.txt
Django==3.1.3
djangorestframework==3.12.2
PyJWT==v1.7.1
mysqlclient==2.0.1
django-mysql==3.9
django-cors-headers==3.5.0
djoser==2.1.0
djangorestframework_simplejwt==4.3.0
markdown==3.3.7
django-filter==21.1
pillow==9.1.1
